Changed Witch's Path description to not be identical with that of the Iron Cities
I changed the description of "Witch's Path", which was identical to that
of the Iron Cities (and indeed, refered to them). I wrote this new
description in a similar way to how the other descriptions were written;
I thought that it would at least be better than having the Iron Cities
description for the Witch's Path.