add "final response phase" flag, for before end turn, combat
This is used to signal when it is still a "response" but the stack
is empty. Intuitively, it should be like the "end turn" and "combat"
events go on the stack, and the opponent should have a chance to
respond to them. (It's not exactly like that, in that if the
opponent does play a response, then when the stack is resolved the
first player gets a chance to play more cards at normal speed.)