add some corrective code for turn announcement text
I'm not sure exactly why this is necessary when it wasn't before,
but I find that when I don't have this, in games vs AI it often
announces the AI's turn as my turn, when I have the opportunity
to respond to the AI's end of play.

This seems like it should only be necessary if the "new turn"
message essentially coincides with the "end turn" message. In MP
games, I guess I would expect that since the new turn always
begins with a proper phase (not a response phase), that this code
that I added would not have any effect.