disable "trampling move" code
since it seems unused (and breaks yogi)

the reasons that it breaks the yogi are a bit complicated. the
issue is, when a unit attacks and kills another unit, it is
supposed to move and take its place and become exhausted. however,
currently exhaustion effects are applied when the combat is
resolved. exhausted units should not be able to move. to work
around this I guess at some point a temporary status effect was
added "trampling" which is added when a unit wins a fight, and
a trampling move phase is added just after combat which moves all
the units with trampling and removes the trampling modifier.

it seems to work fine though if we just ignore the trampling
modifier and move them normally. if exhaustion is not triggered
by combat and only by moving and using abilities, then it works
fine -- the units become exhausted naturally when they reach the
extent of their moves.

the reason it was breaking the yogi is that sleeping was is
implemented by implying exhaustion, but because the trampling move
code ignores exhaustion, even the units that killed the yogi would
move after.