move the fallback font to a different filename from the "real" font
this helps with some issues about git being confused -- it knows
that it is supposed to ignore this file, but it is also committed
to the repo, so normally it will track changes to the changed
content. sometimes strange things happen when i pull from upstream.

it also makes it easier to tell people how to install core mellow
"just drop these files into data/fonts/:
http://theargentlark.com/david/core-mellow.zip"
and nothing is being overwritten in this process