Revert `Diseased Corpse` particle system.
The `smoke_candidate` particle system is intended to be a heavy fire
and smoke particle system for an expansion set card intended to put a
land on fire. It is WIP yet, and (thanks to Dave's advice) is
currently missing a fade out for the smoke.

Revert `Diseased Corpse` to the `disease` particle system.