Reorder starter decks.
Moves Entropia decks to the high line and Materia decks to the low
line. The idea is to have the associative area of decks with the 'most
magic/alien schools' in the higher line, with an association with
spirit; and have the 'most mundane schools' in the lower line, with an
association with ground, stones and nature. As a good side effect the
monocolor Gaea decks is now completely opposite to the monocolor
Entropia deck.