Defer and enlarge clickable widget unhighlight.
Defer and enlarge in time the clickable widget unhighlighting. It's
hard to explain why, but I think it is more visually pleasant, more
smooth. You can try to catch the difference by reverting this commit.

I want to do the same to the 'edit deck / new deck' facility, but
because those are cairo canvases, it will take me a lot more time and
lines to do it there.