Forbid `debug_*_enemy(...)`.
These are working on the player who is not currently in turn. The name
of the function suggests they work on the player not controlled by the
current client.

In other words. With current implementation, in a multiplayer game
these can be used to inject debugging commands to the player
controlled by _this client_, when used during the turn of the opposing
player. Also they can be used to inject debugging commands to the
player controlled by _the opposing client_, when used during _my_
turn. The latter use is not cheating, as it'd be giving an advantage
to the opponent; but the former use is cheating, as it'd be giving a
mischievously advantage to "the _me_".