Fix a tiles layout problem in a puzzle.
I guess this puzzle's layout corresponds to a previous version of the
game, that would have used different functions for arranging the tiles
through the match window space in a different way.

These changes make every opponent seal clearly visible. Seals are not
the point in this puzzle. because it is won by reducing the opponent's
health points to zero by using only damage spells. But that does not
justify leaving the layout _broken_, I think.