Shop page restructure, changed behavior of bougth packs
If player owns a pack, the option to open it no longer replaces the rest
of the Daily Sale but instead of that will shop up in the right bottom
corner.

I split the code into more functions to make it more modular. This
should make integrating different packs more clear.

This commit contains a bug where the label with number of owned packs
doesn't update correctly. It shows the old value (You had 6 packs,
bought 1, the label shows 6. You buy another one, the label shows 7.)