buy_pack function update
I changed the code to look like the buy_card function to make it more
consistent.

The main thing is that the data (new pack, price for the pack) were
updated by the server and that was causing a delay. This problem is
solved by adding add function and updating the account data immediately.