Proper colorless icon in the deck preview widget.
This is super nasty because it leverages some Anura specific knowledge
actually beyond the `citadel` module, but well... at least it works
for me, maybe it works for everybody.

Replace the whitened hex introduced by the previous commit `76fee6ff3`
for a proper _neutral_ colorless icon as seen in the Library.

This rev, and the previous rev, together fix issue 230 of the issue
tracker:

https://github.com/davewx7/citadel/issues/230